docs: comenta y refactoriza monitoreo de camiones
