+++ apps/mfe-dashboard/src/modules/monitoreo-camiones/mocks/mockCamiones.ts
@@ -1,0 +1,167 @@
+import type { CamionHu11, EstadoCamionHu11 } from "@nanutech/api-client";
+
+// Datos de respaldo: se usan cuando el backend no responde o durante desarrollo local.
+export const mockCamiones: CamionHu11[] = [
+  {
+    id: "unidad-001",
+    placa: "ABC-123",
+    marca: "Volvo",
+    modelo: "FH16",
+    anio: 2022,
+    capacidad_ton: 28,
+    estado: "EN_JORNADA",
+    gps_habilitado: true,
+    vin: "VINABC123000001",
+    color: "Azul",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 25495,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: "2026-03-12",
+    proxima_fecha_mantenimiento: "2026-06-12",
+    horas_movimiento: 47.3,
+    horas_detenido: 47.3,
+    horas_totales: 94.5,
+    kilometros_totales: 25495,
+    ultimo_gps_at: "2026-05-05T05:13:00Z",
+    activo: true,
+  },
+  {
+    id: "unidad-002",
+    placa: "DEF-456",
+    marca: "Scania",
+    modelo: "R450",
+    anio: 2023,
+    capacidad_ton: 30,
+    estado: "EN_JORNADA",
+    gps_habilitado: true,
+    vin: "VINDEF456000002",
+    color: "Blanco",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 18310,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: "2026-03-04",
+    proxima_fecha_mantenimiento: "2026-06-04",
+    horas_movimiento: 72,
+    horas_detenido: 0,
+    horas_totales: 72,
+    kilometros_totales: 18310,
+    ultimo_gps_at: "2026-05-05T05:12:00Z",
+    activo: true,
+  },
+  {
+    id: "unidad-003",
+    placa: "GHI-789",
+    marca: "Mercedes-Benz",
+    modelo: "Actros",
+    anio: 2021,
+    capacidad_ton: 26,
+    estado: "EN_JORNADA",
+    gps_habilitado: true,
+    vin: "VINGHI789000003",
+    color: "Azul",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 31500,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: "2026-02-22",
+    proxima_fecha_mantenimiento: "2026-05-22",
+    horas_movimiento: 27.5,
+    horas_detenido: 0,
+    horas_totales: 27.5,
+    kilometros_totales: 31500,
+    ultimo_gps_at: "2026-05-05T05:14:00Z",
+    activo: true,
+  },
+  {
+    id: "unidad-004",
+    placa: "JKL-012",
+    marca: "Volvo",
+    modelo: "FMX",
+    anio: 2020,
+    capacidad_ton: 24,
+    estado: "EN_JORNADA",
+    gps_habilitado: true,
+    vin: "VINJKL012000004",
+    color: "Rojo",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 20440,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: "2026-01-18",
+    proxima_fecha_mantenimiento: "2026-05-18",
+    horas_movimiento: 42,
+    horas_detenido: 0,
+    horas_totales: 42,
+    kilometros_totales: 20440,
+    ultimo_gps_at: "2026-05-05T05:11:00Z",
+    activo: true,
+  },
+  ...["MNO-345", "PQR-678", "STU-901", "VWX-234"].map((placa, index) => ({
+    id: `unidad-disponible-${index + 1}`,
+    placa,
+    marca: index % 2 === 0 ? "Scania" : "Volvo",
+    modelo: index % 2 === 0 ? "P360" : "FH",
+    anio: 2020 + index,
+    capacidad_ton: 24 + index,
+    estado: "DISPONIBLE" as EstadoCamionHu11,
+    gps_habilitado: index < 2,
+    vin: `VINDISP${index + 1}000000`,
+    color: "Blanco",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 0,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: null,
+    proxima_fecha_mantenimiento: null,
+    horas_movimiento: 0,
+    horas_detenido: 0,
+    horas_totales: 0,
+    kilometros_totales: 0,
+    ultimo_gps_at: null,
+    activo: true,
+  })),
+  ...["YZA-567", "BCD-890"].map((placa, index) => ({
+    id: `unidad-mantenimiento-${index + 1}`,
+    placa,
+    marca: "Mercedes-Benz",
+    modelo: "Arocs",
+    anio: 2019 + index,
+    capacidad_ton: 22,
+    estado: "MANTENIMIENTO" as EstadoCamionHu11,
+    gps_habilitado: false,
+    vin: `VINMANT${index + 1}000000`,
+    color: "Gris",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 0,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: "2026-04-20",
+    proxima_fecha_mantenimiento: "2026-05-20",
+    horas_movimiento: 0,
+    horas_detenido: 0,
+    horas_totales: 0,
+    kilometros_totales: 0,
+    ultimo_gps_at: null,
+    activo: true,
+  })),
+  ...["ZZZ-111", "ZZZ-222", "ZZZ-333"].map((placa, index) => ({
+    id: `unidad-en-uso-${index + 1}`,
+    placa,
+    marca: "Volvo",
+    modelo: "VM",
+    anio: 2020,
+    capacidad_ton: 18,
+    estado: "EN_JORNADA" as EstadoCamionHu11,
+    gps_habilitado: false,
+    vin: `VINUSO${index + 1}000000`,
+    color: "Negro",
+    tipo_combustible: "DIESEL",
+    kilometraje_actual: 0,
+    fecha_registro: "2026-04-07",
+    ultima_fecha_mantenimiento: null,
+    proxima_fecha_mantenimiento: null,
+    horas_movimiento: 0,
+    horas_detenido: 0,
+    horas_totales: 0,
+    kilometros_totales: 0,
+    ultimo_gps_at: null,
+    activo: true,
+  })),
+];
+
