+++ apps/mfe-dashboard/vite.config.ts
@@ -44,9 +44,9 @@ export default defineConfig({
       include: [
         "src/App.tsx",
         "src/modules/dashboard-admin/**/*.{ts,tsx}",
-        "src/modules/monitoreo-camiones/*/.{ts,tsx}",
+        "src/modules/monitoreo-camiones/**/*.{ts,tsx}",
       ],
-      exclude: ["src/main.tsx", "src/**/*.d.ts", "src/**/index.ts"],
+      exclude: ["src/main.tsx", "src/**/*.d.ts", "src/**/index.ts", "src/**/types.ts"],
     },
   },
 });
